backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Django Backend"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = 'kernelCI' }]

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.11"
djangorestframework = "^3.15.2"
gunicorn = "^22.0.0"
psycopg = "^3.1.19"
django-query-builder = "^3.2.0"
django-cors-headers = "^4.4.0"
beautifulsoup4 = "^4.12.3"
requests = "^2.32.3"
pydantic = "^2.10.3"
drf-spectacular = "^0.28.0"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
mypy = "^1.10.0"
flake8 = "^7.1.0"
pytest = "^8.3.4"
pytest-django = "^4.10.0"
pep8-naming = "^0.14.1"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "kernelCI.settings"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
